Rename and move class up one namespace
